+++ src/fstcore/integer64/integer64_v11.h
@@ -40,8 +40,9 @@
 #include <ostream>
 
 
-void fdsWriteInt64Vec_v11(std::ofstream &myfile, long long* int64Vector, unsigned int nrOfRows, unsigned int compression, std::string annotation);
+void fdsWriteInt64Vec_v11(std::ofstream &myfile, long long* int64Vector, unsigned long long nrOfRows, unsigned int compression, std::string annotation);
 
-void fdsReadInt64Vec_v11(std::istream &myfile, long long* int64Vector, unsigned long long blockPos, unsigned int startRow, unsigned int length, unsigned int size);
+void fdsReadInt64Vec_v11(std::istream &myfile, long long* int64Vector, unsigned long long blockPos, unsigned long long startRow,
+  unsigned long long length, unsigned long long size);
 
 #endif // INT64_V11_H
